+++ README.md
@@ -34,3 +34,14 @@ npm run build
 ```
 
 > You can preview the built app with `npm run preview`, regardless of whether you installed an adapter. This should _not_ be used to serve your app in production.
+
+
+# Todo:
+
+* Add graphics to error page
+* Make dato ranges automatic
+* Integrate sales to Fiken
+* Proper logging
+* Database?
+* TS-ify things
+
